requirements.txt
flask
flask-cors
sentence-transformers
faiss-cpu
transformers
torch
pymupdf
numpy